Another version of createDataRunDefinitions.py was already present.
Updated that version instead.
